bug fixed:
a bug prevented the correct plotting of several files each containing multiple blocks of data.
An identical name of the tmp files used to dump each block of data was used when handling each of the
data files (output.1, output.2) so that new data blocks overwrote the old one, causing the disappeareance
of the latter: in practice only on data file was plotted.

Fixed by adding a random integer number to naming convention.