handle function of single columns as an argument of the -u option see help message for more details.
It is now possible to apply on the fly some function defined in XMGRACE
to a single column by specifying that in the -u option, e.g.
plot -u1:'cos($2)' file.dat

This is achieved by analyzing the argument of -u option: if that is not a number
the corresponding operation is added to the batch file and used to plot the
n-th data block.

This is still EXPERIMENTAL and should not be used for multiple data sets or files.