Version v2.1:
- fixed a bug with the -c<C1>:<C2> option to generate a gradient of colors.
In the previous version a specification of the color label to use in the BATCH
file was missing. As a result the colors of the line ranged from 1 to 15, i.e. the
number of colors available in the passed configuration file $HOME/.grace/gracerc.user
The bug has been solved by adding a statement to write the line color (2+line_number)
to the BATCH script.