v2.2
New feature: set the graph type with -st option. -st[xy][xydx][xydy][xydxdy]
correspondiog to graph with no error bars, error bars alon X, Y or both directions.

Number of columns will be adjusted by default or if passed with the -u option a check
will prevent errors.

This is a global specification.
A better implementation would be based on the column syntax and passed silently to the script.
For example using the -u1:2:4 option to use the xydx type or the -u1:2::4 to use the xydy or again
the -u1:2:4:3 to use xydxdy. I will work on that.